requirements.txt
# Backend - API e Processamento
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Processamento de PDFs
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0

# IA para Chatbot (escolha uma das opções)
openai==1.3.7  # Se for usar OpenAI
# google-generativeai==0.3.1  # Ou Google Gemini (gratuito)

# Utilidades
python-dotenv==1.0.0
python-multipart==0.0.6

# CORS para comunicação frontend-backend
fastapi-cors==0.0.6

# Banco de dados simples (opcional)
tinydb==4.8.0